pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-campus-inspection"
version = "0.1.0"
description = "Campus infrastructure inspection perception layer powered by YOLOv8"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
ai-inspect = "main:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["perception*"]

